phase4/requirement.txt
python-dotenv==0.14.0
tabulate==0.8.7
PyMySQL==0.10.1